remove the old planning approach
