+++ tsl/src/nodes/vector_agg/CMakeLists.txt
@@ -1,2 +1,3 @@
-set(SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/vector_agg.c)
+set(SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/exec.c
+            ${CMAKE_CURRENT_SOURCE_DIR}/plan.c)
 target_sources(${TSL_LIBRARY_NAME} PRIVATE ${SOURCES})
