+++ tsl/src/nodes/vector_agg/exec.c
@@ -1,0 +1,72 @@
+/*
+ * This file and its contents are licensed under the Timescale License.
+ * Please see the included NOTICE for copyright information and
+ * LICENSE-TIMESCALE for a copy of the license.
+ */
+
+#include <postgres.h>
+
+#include <commands/explain.h>
+#include <executor/executor.h>
+#include <nodes/extensible.h>
+#include <nodes/makefuncs.h>
+#include <nodes/nodeFuncs.h>
+
+#include "exec.h"
+
+#include "nodes/decompress_chunk/compressed_batch.h"
+#include "nodes/decompress_chunk/exec.h"
+
+static void
+vector_agg_begin(CustomScanState *node, EState *estate, int eflags)
+{
+	CustomScan *cscan = castNode(CustomScan, node->ss.ps.plan);
+	node->custom_ps =
+		lappend(node->custom_ps, ExecInitNode(linitial(cscan->custom_plans), estate, eflags));
+}
+
+static void
+vector_agg_end(CustomScanState *node)
+{
+	ExecEndNode(linitial(node->custom_ps));
+}
+
+static void
+vector_agg_rescan(CustomScanState *node)
+{
+	if (node->ss.ps.chgParam != NULL)
+		UpdateChangedParamSet(linitial(node->custom_ps), node->ss.ps.chgParam);
+
+	ExecReScan(linitial(node->custom_ps));
+}
+
+static TupleTableSlot *
+vector_agg_exec(CustomScanState *node)
+{
+	// return ExecProcNode(linitial(node->custom_ps));
+	DecompressChunkState *ds = (DecompressChunkState *) linitial(node->custom_ps);
+	return decompress_chunk_exec_vector_agg_impl(node, ds);
+}
+
+static void
+vector_agg_explain(CustomScanState *node, List *ancestors, ExplainState *es)
+{
+	/* noop? */
+}
+
+static struct CustomExecMethods exec_methods = {
+	.CustomName = "VectorAgg",
+	.BeginCustomScan = vector_agg_begin,
+	.ExecCustomScan = vector_agg_exec,
+	.EndCustomScan = vector_agg_end,
+	.ReScanCustomScan = vector_agg_rescan,
+	.ExplainCustomScan = vector_agg_explain,
+};
+
+Node *
+vector_agg_state_create(CustomScan *cscan)
+{
+	CustomScanState *state = makeNode(CustomScanState);
+	state->methods = &exec_methods;
+	return (Node *) state;
+}
