+++ tsl/src/nodes/vector_agg/exec.h
@@ -6,7 +6,6 @@
 
 #include <postgres.h>
 
-#include <nodes/plannodes.h>
 #include <nodes/execnodes.h>
 
 typedef struct VectorAggState
@@ -14,13 +13,4 @@ typedef struct VectorAggState
 	CustomScanState custom;
 } VectorAggState;
 
-typedef struct VectorAggPlan
-{
-	CustomScan custom;
-} VectorAggPlan;
-
-extern Plan *vector_agg_plan_create(Agg *agg, CustomScan *decompress_chunk);
-
 extern Node *vector_agg_state_create(CustomScan *cscan);
-
-extern void _vector_agg_init(void);
