--- tsl/src/partialize_agg.h
@@ -1,11 +1,0 @@
-/*
- * This file and its contents are licensed under the Timescale License.
- * Please see the included NOTICE for copyright information and
- * LICENSE-TIMESCALE for a copy of the license.
- */
-#pragma once
-
-extern bool apply_vectorized_agg_optimization(PlannerInfo *root, AggPath *aggregation_path,
-											  Path *subpath);
-
-extern void tsl_postprocess_plan(PlannedStmt *stmt);
