+++ tsl/src/planner.h
@@ -17,3 +17,4 @@ void tsl_set_rel_pathlist_query(PlannerInfo *, RelOptInfo *, Index, RangeTblEntr
 void tsl_set_rel_pathlist_dml(PlannerInfo *, RelOptInfo *, Index, RangeTblEntry *, Hypertable *);
 void tsl_set_rel_pathlist(PlannerInfo *root, RelOptInfo *rel, Index rti, RangeTblEntry *rte);
 void tsl_preprocess_query(Query *parse);
+void tsl_postprocess_plan(PlannedStmt *stmt);
